The complite controller "AddTask"

+++ src/main/java/pl/mloza/repozytory/ProjectRepository.java
@@ -3,4 +3,4 @@ package pl.mloza.repozytory;
 import org.springframework.data.repository.CrudRepository;
 import pl.mloza.entity.Project;
 
-public interface ProjectRepository extends CrudRepository<Project,Long> { }
+public interface ProjectRepository extends CrudRepository<Project,Integer> { }
